# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Notifier API",
        "description": "API for managing orders with notifications",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/devices": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Register device token",
                "description": "Save or update FCM device token. Deduplicates per user+token. Requires devices:write ability.",
                "operationId": "88857646312329caabd3f773d26b6376",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "platform"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "fcm_token_here_abc123xyz..."
                                    },
                                    "platform": {
                                        "type": "string",
                                        "enum": [
                                            "android",
                                            "ios",
                                            "web"
                                        ],
                                        "example": "android"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device token registered/updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "device_token": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "token": {
                                                    "type": "string"
                                                },
                                                "platform": {
                                                    "type": "string",
                                                    "example": "android"
                                                },
                                                "last_seen_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - missing devices:write ability"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/devices/{id}": {
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Remove device token",
                "description": "Delete a device token. User can only delete their own tokens. Requires devices:write ability.",
                "operationId": "19e2a79d24664fc47fdb47c8c153aa1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device token ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device token deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device token deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - not your token or missing devices:write ability"
                    },
                    "404": {
                        "description": "Device token not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}/notify": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Manually trigger push notification",
                "description": "Send FCM push notification for order to all user devices. Requires notify:send ability.",
                "operationId": "7629f58aeaa5b650b518881dce47e2fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notifications sent"
                                        },
                                        "result": {
                                            "properties": {
                                                "success": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "failed": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No device tokens found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - not your order or missing notify:send ability"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}